chore: update golangci-lint plugins and enable unconvert linter (#3160)
## Changes

- Bump `kube-api-linter` plugin to latest commit
- Add `go: '1.27'` toolchain pin and enable `unconvert` linter in
`.golangci.yaml`
- Set `GOTOOLCHAIN` env var when building custom golangci-lint binary in
Makefile
- Bump `tools/go.mod` to Go 1.27
- Remove redundant type conversions flagged by `unconvert` (e.g.
unnecessary `string()`, `int32()` casts) across 17 files

+++ controller/pkg/agentgateway/remotehttp/connection_resolver.go
@@ -195,7 +195,7 @@ func (r *defaultResolver) resolveTunnelProxy(
 		}
 		var port int32
 		if p := ptr.OrEmpty(backendRef.Port); p != 0 {
-			port = int32(p)
+			port = p
 		} else if backend.Static.Port != 0 {
 			port = backend.Static.Port
 		} else {
